requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==2.2
django-allauth==0.38.0
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.6
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
